Updated readme to point to the most helpful wiki page
The previous version of the front page required hunting around to find the right wiki page for the user guide.  I called out the correct User Guide page with a "Getting Started" section.

I also edited the link in the Meta section.  It seemed to point to a page that was just a handful of IDE and object mapping tips, so I updated the URL to drop the visitor onto the top level wiki page.  I'll propose some additional wiki cleanups in future commits.